face-monitoring-master/server/requirements.txt
asn1crypto==0.24.0
click==7.1.2
cryptography==2.1.4
enum34==1.1.6
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SocketIO==4.3.1
gyp==0.1
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
numpy==1.16.6
opencv-python==4.2.0.32
pandas==0.24.2
pycrypto==2.6.1
pygobject==3.26.1
python-dateutil==2.8.1
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2020.1
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
supervisor==4.2.0
Werkzeug==1.0.1
